EES-7374 - stop pushing the data file list up to ReleaseDataPage
ReleaseDataUploadsSection pushed its query results up to ReleaseDataPage
through an onDataFilesChange prop in an effect, purely so the page could
count COMPLETE files and use that as a `key` to remount two lazy tabs.

The page can read the same query itself. It shares a query key with the
section, so it is served from the cache without an extra request, and the
prop, the effect and the page's useState all go.

Also in passing:

- Collapse the `refetchDataFiles(); refetchDataSetUploads();` pairs in
  handleSubmit, handleDeleteUploadConfirm and handleDataSetImport into a
  single refreshDataFileLists callback that runs them concurrently.
- Drop the effect in DataFilesReorderableList that re-seeded local state
  from its prop. The list is unmounted whenever reordering stops, so the
  useState initialiser is enough, and the effect meant a background list
  refetch mid-drag would discard the user's in-progress ordering.
- Memoise the two upload filters, for consistency with the two directly
  below them.
- Move the static "Before you start" guidance into its own component.
